Improve error message for a bad hostname (#2523)
Closes #1941

When using the `--hostname` argument with a name that isn't supported by
the OS the `SocketException` would have a null `address` and `port`
field. In this case we'd get a stack trace trying to emit an error which
makes it look like a bug in `build_runner` rather than a usage error.

Since we don't know all the conditions that lead to a `SocketException`
the message is somewhat vague.

+++ build_runner/lib/src/entrypoint/serve.dart
@@ -72,10 +72,14 @@ class ServeCommand extends WatchCommand {
       }));
     } on SocketException catch (e) {
       var listener = Logger.root.onRecord.listen(stdIOLogListener());
-      logger.severe(
-          'Error starting server at ${e.address.address}:${e.port}, address '
-          'is already in use. Please kill the server running on that port or '
-          'serve on a different port and restart this process.');
+      if (e.address != null && e.port != null) {
+        logger.severe(
+            'Error starting server at ${e.address.address}:${e.port}, address '
+            'is already in use. Please kill the server running on that port or '
+            'serve on a different port and restart this process.');
+      } else {
+        logger.severe('Error starting server on ${options.hostName}.');
+      }
       await listener.cancel();
       return ExitCode.osError.code;
     }
